--- a/recipe/grpo/grpo_ray_trainer.py
+++ b/recipe/grpo/grpo_ray_trainer.py
@@ -107,15 +107,9 @@ class RayGRPOTrainer(RayPPOTrainer):
                         if not self.async_rollout_mode:
                             gen_batch_output = self.actor_rollout_wg.generate_sequences(gen_batch)
                         else:
-                            t1 = time.time()
                             self.async_rollout_manager.wake_up()
-                            t2 = time.time()
                             gen_batch_output = self.async_rollout_manager.generate_sequences(gen_batch)
-                            t3 = time.time()
                             self.async_rollout_manager.sleep()
-                            t4 = time.time()
-                            print(f"engine wake_up:{t2 - t1}, generate:{t3 - t2}, sleep:{t4 - t3}s")
-                            
 
                     if self.config.algorithm.adv_estimator == AdvantageEstimator.REMAX:
                         with _timer("gen_max", timing_raw):