--- a/recipe/moe/moe_trainer/dump.py
+++ b/recipe/moe/moe_trainer/dump.py
@@ -1,0 +1,1 @@
+/cpfs/user/guangsu/verl-debug/dump.py